requirements.txt
pymongo
requests
pandas
pillow